Set up channel receivers early (#145)
At present, components can race with each other and try to send a
message on a channel before the other side has managed to set up
their receiver. When this happens, an error is raised, which can lead
to process termination.

Instead, set up all the receivers at the initialization stage (e.g.
when calling new()). This requires some trickery to pass the mutable
receiver down into the respective start() -> receive() methods, which
is achieved with an option inside a cell. When the mutable receiver is
consumed, it is replaced with a None value.

The stop methods of the various structs that need this all consume the
reference (i.e. take &mut self), so there's no chance of the start()
method being called a second time (leading to the `expect` calls being
triggered).